chore: dump font icons as json instead of csv
