+++ src/ObjCommon/Game/T6/FontIcon/FontIconCommonT6.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+
+namespace T6::font_icon
+{
+    std::string GetJsonFileNameForAssetName(const std::string& assetName);
+}
