+++ src/ObjWriting/Game/T6/FontIcon/DumperFontIconJsonT6.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "Dumping/AbstractAssetDumper.h"
+#include "Game/T6/T6.h"
+
+namespace T6
+{
+    void DumpFontIconAsJson(const AssetDumpingContext& context, const FontIcon& fontIcon);
+}
